requirements.txt
# ================================================
# CORE DEPENDENCIES - Minimal install for basic operation
# Total size: ~500MB (no PyTorch)
# ================================================

# Core Agent Engine
anthropic>=0.40.0
litellm>=1.57.0
python-dotenv>=1.0.0
aiofiles>=23.2.1

# Brain - Vector Database (LanceDB — ACID-compliant, crash-safe)
lancedb>=0.6.0
sentence-transformers>=2.2.0
pyarrow>=14.0.0
pydantic-settings>=2.0.0

# Monitoring & Notifications
python-telegram-bot>=21.0

# Web Dashboard
fastapi>=0.109.0
uvicorn>=0.25.0
websockets>=12.0

# Security & Auto-Update
pip-audit>=2.6.1
safety>=3.0.1
cryptography>=41.0.7

# Web Tools
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
requests>=2.31.0
tavily-python>=0.3.0       # AI-optimised search API (set TAVILY_API_KEY in .env)
duckduckgo-search>=6.0.0   # No-key fallback (may be rate-limited on cloud IPs)

# Utilities
pydantic>=2.5.3
pyyaml>=6.0.1
psutil>=5.9.7
structlog>=24.1.0

# Social Media Tools (X/Twitter OAuth 1.0a)
requests-oauthlib>=1.3.1

# Email & Calendar Tools (Gmail, Outlook, iCloud support)
caldav>=1.3.6
icalendar>=5.0.11
pytz>=2023.3

# ================================================
# OPTIONAL DEPENDENCIES - Install separately if needed
# ================================================
# For multi-channel support:
#   pip install discord.py slack-bolt slack-sdk
#
# For full browser automation:
#   pip install playwright && playwright install chromium
#
# For computer use automation:
#   pip install pyautogui mss
#
# For development:
#   pip install pytest pytest-asyncio black ruff
twilio>=9.0.0
python-multipart>=0.0.6
playwright>=1.41.0

# For intent classification with DistilBERT (install on EC2)
transformers
torch